add basic delivery status implementation

+++ queue/src/queue_info.rs
@@ -172,12 +172,12 @@ impl Alert {
     }
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct PushStatus {
     pub subscriber_name: String,
     pub retries_remaining: u32,
     pub tries: u32,
-	pub status_code: Option<u32>,
+	pub status_code: Option<u16>,
 	pub url: String,
     pub msg: Option<String>
 }
